[v1.15] backport: google.protobuf.UInt64Value does not marshal correctly in CRDs (#8587)
* fix: google.protobuf.UInt64Value does not marshal correctly in CRDs (#8570)

* fix: google.protobuf.UInt64Value does not marshal correctly in kubernetes CRDs

* Add validation tests in helm upgrade

* More descriptive changelog

* update test to use a field which can be backported

* Fix double colon

* revert test changes

* bump to fixed version

* Test reduced max

* Adding changelog file to new location

* Deleting changelog file from old location

* Change type to Uint32

* Adding changelog file to new location

* Deleting changelog file from old location

* Fix tests

* Update test comment


* Update changelog

* Fix changelog

---------


* Update changelog

* Fix changelog location

* Add skip directives

* ci: Fix check skip directives script

---------

+++ projects/gloo/pkg/api/v1/enterprise/options/caching/caching.pb.clone.go
@@ -67,9 +67,9 @@ func (m *Settings) Clone() proto.Message {
 	}
 
 	if h, ok := interface{}(m.GetMaxPayloadSize()).(clone.Cloner); ok {
-		target.MaxPayloadSize = h.Clone().(*github_com_golang_protobuf_ptypes_wrappers.UInt64Value)
+		target.MaxPayloadSize = h.Clone().(*github_com_golang_protobuf_ptypes_wrappers.UInt32Value)
 	} else {
-		target.MaxPayloadSize = proto.Clone(m.GetMaxPayloadSize()).(*github_com_golang_protobuf_ptypes_wrappers.UInt64Value)
+		target.MaxPayloadSize = proto.Clone(m.GetMaxPayloadSize()).(*github_com_golang_protobuf_ptypes_wrappers.UInt32Value)
 	}
 
 	return target
